+++ bin-proto/src/types/mod.rs
@@ -1,15 +1,11 @@
 //! Utility types
 
 mod array;
-mod char;
-/// Definitions for the `std::collections` module.
 mod collections;
 mod cstring;
 mod marker;
 mod numerics;
 mod option;
-mod range;
-/// Definitions for smart pointers in the `std` module.
 mod smart_ptr;
 mod string;
 mod tuple;